+++ src/components/Layout/index.tsx
@@ -1,5 +1,6 @@
 import { FC, PropsWithChildren } from "react"
 import getThemedClassName from "~/themes/getThemedClassName"
+import { Theme } from "~/themes/Theme"
 import ThemeContext from "~/themes/ThemeContext"
 import CurrentBanner from "../banners/CurrentBanner"
 import Footer, { Props as FooterProps } from "./Footer"
@@ -8,7 +9,7 @@ import styles from "./index.module.scss"
 export interface Props extends PropsWithChildren {
     footerPromos?: FooterProps["promos"]
     headerPromos?: HeaderProps["promos"]
-    theme: "bw" | "day" | "kc" | "night"
+    theme: Theme
 }
 const Layout: FC<Props> = ({ children, headerPromos, footerPromos, theme }) => {
     const className = getThemedClassName(theme, styles, "layout")