+++ src/themes/Theme.d.ts
@@ -1,1 +1,1 @@
-export type Theme = "bw" | "day" | "kc" | "night"
+export type Theme = "bw" | "day" | "kc" | "night" | "webcomic"