+++ src/themes/useBackgroundColor.ts
@@ -6,6 +6,7 @@ const COLORS: Readonly<Record<Theme, ThemeColor>> = {
     day: "white",
     kc: "cream",
     night: "black",
+    webcomic: "white"
 }
 const useBackgroundColor = () => COLORS[useTheme()]
 export default useBackgroundColor