+++ CHANGELOG.md
@@ -9,7 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+-   _Parry 'n' Carney: Friends for Life_.
+
 ### Changed
+
+-   Updated _PhyloPic_ link to `www.phylopic.org`.
 
 ### Deprecated